requirements.txt
torch
torchaudio

transformers
accelerate
datasets

librosa
soundfile

lingua-language-detector
